添加查询会员加入待审批记录

+++ src/main/java/com/tw/clubmanagement/entity/BaseEntity.java
@@ -18,12 +18,12 @@ public class BaseEntity {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   protected Integer id;
 
-  private long createdBy;
+  private int createdBy;
 
   @CreationTimestamp
   private Date createdAt;
 
-  private long updatedBy;
+  private int updatedBy;
 
   @UpdateTimestamp
   private Date updatedAt;
